+++ changelog.md
@@ -4,6 +4,7 @@
 ### fix
 - (#897) stucked arrival notifications.
 - showing native notifications preview with the seelen ones (now only seelen arrival will be shown).
+- (#934) invisible edge window on dock.
 
 ## [2.3.7]
 ### enhancements